+++ lib/src/buttons/button.dart
@@ -1,5 +1,8 @@
 import 'package:arna/arna.dart';
 
+/// Button types.
+enum ButtonType { normal, colored, destructive, suggested }
+
 class ArnaButton extends StatelessWidget {
   /// Creates a button.
   const ArnaButton({
@@ -8,6 +11,7 @@ class ArnaButton extends StatelessWidget {
     this.icon,
     required this.onPressed,
     this.tooltipMessage,
+    this.buttonType = ButtonType.normal,
     this.isFocusable = true,
     this.autofocus = false,
     this.accentColor,
@@ -27,6 +31,9 @@ class ArnaButton extends StatelessWidget {
   /// The tooltip message of the button.
   final String? tooltipMessage;
 
+  /// The type of the button.
+  final ButtonType buttonType;
+
   /// Whether this button is focusable or not.
   final bool isFocusable;
 
@@ -44,14 +51,18 @@ class ArnaButton extends StatelessWidget {
   /// The semantic label of the button.
   final String? semanticLabel;
 
-  Widget _buildChild(BuildContext context, bool enabled) {
+  Widget _buildChild(BuildContext context, bool enabled, Color accent) {
     final List<Widget> children = [];
     if (icon != null) {
       Widget iconWidget = Icon(
         icon!,
         size: Styles.iconSize,
         color: ArnaDynamicColor.resolve(
-          !enabled ? ArnaColors.disabledColor : ArnaColors.iconColor,
+          !enabled
+              ? ArnaColors.disabledColor
+              : buttonType == ButtonType.normal
+                  ? ArnaColors.iconColor
+                  : ArnaDynamicColor.luminance(accent, ArnaColors.iconColor),
           context,
         ),
       );
@@ -68,7 +79,12 @@ class ArnaButton extends StatelessWidget {
                 color: ArnaDynamicColor.resolve(
                   !enabled
                       ? ArnaColors.disabledColor
-                      : ArnaColors.primaryTextColor,
+                      : buttonType == ButtonType.normal
+                          ? ArnaColors.primaryTextColor
+                          : ArnaDynamicColor.luminance(
+                              accent,
+                              ArnaColors.primaryTextColor,
+                            ),
                   context,
                 ),
               ),
@@ -84,6 +100,18 @@ class ArnaButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    Color accent;
+    switch (buttonType) {
+      case ButtonType.destructive:
+        accent = ArnaColors.errorColor;
+        break;
+      case ButtonType.suggested:
+        accent = ArnaColors.accentColor;
+        break;
+      default:
+        accent = accentColor ?? ArnaTheme.of(context).accentColor;
+    }
+
     return Padding(
       padding: Styles.small,
       child: ArnaBaseButton(
@@ -96,25 +124,56 @@ class ArnaButton extends StatelessWidget {
             decoration: BoxDecoration(
               borderRadius: Styles.borderRadius,
               border: Border.all(
-                color: buttonBorder(
+                color: ArnaDynamicColor.resolve(
+                  buttonType == ButtonType.normal
+                      ? !enabled
+                          ? ArnaColors.borderColor
+                          : focused
+                              ? accent
+                              : ArnaColors.borderColor
+                      : !enabled
+                          ? ArnaDynamicColor.colorBlender(
+                              accent,
+                              42,
+                              isBorder: true,
+                            )
+                          : focused
+                              ? ArnaDynamicColor.colorBlender(
+                                  accent,
+                                  28,
+                                  isBorder: true,
+                                )
+                              : ArnaDynamicColor.colorBlender(
+                                  accent,
+                                  42,
+                                  isBorder: true,
+                                ),
                   context,
-                  enabled,
-                  focused,
-                  accentColor ?? ArnaTheme.of(context).accentColor,
                 ),
               ),
-              color: buttonBackground(
+              color: ArnaDynamicColor.resolve(
+                !enabled
+                    ? ArnaColors.backgroundColor
+                    : buttonType == ButtonType.normal
+                        ? pressed
+                            ? ArnaColors.buttonPressedColor
+                            : hover
+                                ? ArnaColors.buttonHoverColor
+                                : ArnaColors.buttonColor
+                        : pressed
+                            ? ArnaDynamicColor.colorBlender(accent, 42)
+                            : hover
+                                ? ArnaDynamicColor.colorBlender(accent, 28)
+                                : focused
+                                    ? ArnaDynamicColor.colorBlender(accent, 28)
+                                    : accent,
                 context,
-                enabled,
-                hover,
-                focused,
-                pressed,
               ),
             ),
             padding: icon != null
                 ? const EdgeInsets.symmetric(horizontal: Styles.padding - 1)
                 : Styles.largeHorizontal,
-            child: _buildChild(context, enabled),
+            child: _buildChild(context, enabled, accent),
           );
         },
         onPressed: onPressed,