+++ lib/src/helpers.dart
@@ -1,37 +1,5 @@
 import 'package:arna/arna.dart';
 
-Color buttonBorder(
-  BuildContext context,
-  bool enabled,
-  bool focused,
-  Color accentColor,
-) {
-  return !enabled
-      ? ArnaDynamicColor.resolve(ArnaColors.borderColor, context)
-      : focused
-          ? accentColor
-          : ArnaDynamicColor.resolve(ArnaColors.borderColor, context);
-}
-
-Color buttonBackground(
-  BuildContext context,
-  bool enabled,
-  bool hover,
-  bool focused,
-  bool pressed,
-) {
-  return ArnaDynamicColor.resolve(
-    !enabled
-        ? ArnaColors.backgroundColor
-        : pressed
-            ? ArnaColors.buttonPressedColor
-            : hover
-                ? ArnaColors.buttonHoverColor
-                : ArnaColors.buttonColor,
-    context,
-  );
-}
-
 double deviceHeight(context) => MediaQuery.of(context).size.height;
 
 double deviceWidth(context) => MediaQuery.of(context).size.width;